+++ src/ExportEdiJobs/ExportEdiJobsFilters/ExportMethodFilter/ExportMethodFilter.test.js
@@ -1,8 +1,6 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import '@folio/stripes-acq-components/test/jest/__mock__';
-
+import { render } from '@folio/jest-config-stripes/testing-library/react';
 import { runAxeTest } from '@folio/stripes-testing';
+
 import Filter from './ExportMethodFilter';
 
 jest.mock('./useConfigs', () => ({