+++ src/ExportJobs/ExportJobsList/ExportJobsList.test.js
@@ -1,8 +1,5 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-
+import { render } from '@folio/jest-config-stripes/testing-library/react';
 import { runAxeTest } from '@folio/stripes-testing';
-import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import { ExportJobsList } from './ExportJobsList';